requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
altgraph==0.17.2
ansicon==1.89.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
async-generator==1.10
async-timeout==3.0.1
asyncssh==2.12.0
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
backoff==2.1.2
base58==2.1.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blinker==1.4
blis==0.7.7
blockfrost-python==0.3.0
brython==3.10.6
bs4==0.0.1
cachelib==0.4.1
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.3.1
click==8.1.3
click-log==0.4.0
colorama==0.4.4
commonmark==0.9.1
configparser==5.2.0
crayons==0.4.0
cryptography==35.0.0
css-html-js-minify==2.5.5
cymem==2.0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
device-detector==5.0.1
discord.py==1.7.3
distlib==0.3.4
distro==1.7.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.16
enlighten==1.10.2
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.4.0
filetype==1.1.0
Flask==2.0.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-Simple-GeoIP==0.2.4
flask-unittest==0.1.2
Flask-WTF==1.0.1
flit==3.4.0
flit_core==3.4.0
future==0.18.2
future-fstrings==1.2.0
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gevent==21.8.0
greenlet==1.1.2
h11==0.13.0
httpserver==1.1.0
huggingface-hub==0.8.1
humanfriendly==10.0
idna==3.3
imagesize==1.2.0
importlib-metadata==4.12.0
iniconfig==1.1.1
ipaddress==1.0.23
ipfsapi==0.4.4
ipfshttpclient==0.4.10
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
json5==0.9.9
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
langcodes==3.3.0
libpff-python-ratom==20220304
libratom==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mega.py==1.0.8
mistune==0.8.4
multiaddr==0.0.9
multidict==5.1.0
murmurhash==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.0.29
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
outcome==1.1.0
packaging==21.3
pandas==1.3.0
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.3
pathlib==1.0.1
pathy==0.6.1
pbr==5.9.0
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipfreeze==2.0.3
pipreqs==0.4.11
platformdirs==2.4.1
plotly==5.1.0
pluggy==1.0.0
prefixed==0.3.2
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.18.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cord @ git+https://github.com/Pycord-Development/pycord@cd70e292c59b6abec4f8f12a4e5889746d21705e
py3-validate-email==1.0.5
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.4
pycryptodomex==3.15.0
pydantic==1.8.2
pygeoip==0.3.2
Pygments==2.10.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.0.1
python-dateutil==2.8.2
python-discord==1.7.3
pytube==11.0.2
pytz==2021.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.7
PyYAML==6.0
pyzmq==23.2.1
qrcode==7.3.1
ratelim==0.1.6
recommonmark==0.5.0
regex==2022.7.25
requests==2.27.1
rsa==4.7.2
scapy==2.4.5
selenium==4.1.2
Send2Trash==1.8.0
simple-geoip==0.1.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
Sphinx==3.5.4
sphinx-autodoc-typehints==1.12.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.37
srsly==2.4.3
sshtunnel==0.4.0
stack-data==0.4.0
striprtf==0.0.20
tabulate==0.8.9
Telethon==1.23.0
tenacity==5.1.5
terminado==0.15.0
thinc==8.0.17
time-tracker-cli==0.0.1
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.9.6
tomli==1.2.1
tomli_w==0.3.0
torch==1.12.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.1
treelib==1.6.1
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
virtualenv==20.13.0
waitress==2.0.0
wasabi==0.9.1
wcwidth==0.2.5
webdriver-manager==3.5.3
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.1
widgetsnbextension==3.6.1
WMI==1.5.1
wsproto==1.1.0
WTForms==3.0.1
yarg==0.1.9
yarl==1.6.3
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
